1925: CE.rust (#1204) - 无编译器检查

+++ Codes/1925-count-square-sum-triples.rs
@@ -2,7 +2,7 @@
  * @Author: LetMeFly
  * @Date: 2025-12-08 21:36:58
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2025-12-08 21:53:08
+ * @LastEditTime: 2025-12-08 21:55:06
  */
 impl Solution {
     pub fn count_triples(n: i32) -> i32 {
@@ -10,8 +10,8 @@ impl Solution {
         for a in 1..n+1 {
             for b in 1..n+1 {
                 let k = a * a + b * b;
-                let c = k.sqrt();
-                if c as i32 <= n && c * c == k {
+                let c = (k as f64).sqrt();
+                if c <= n && c * c == k {
                     ans += 1;
                 }
             }
